LPD-50383 Encapsulate common logic in Base classes and avoid inherit from specific classes
