+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/display/context/ViewSpaceFilesSectionDisplayContext.java
@@ -10,18 +10,15 @@ import com.liferay.object.service.ObjectDefinitionService;
 import com.liferay.object.service.ObjectDefinitionSettingLocalService;
 import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.Portal;
 
 import jakarta.servlet.http.HttpServletRequest;
-
-import java.util.Map;
 
 /**
  * @author Roberto Díaz
  */
 public class ViewSpaceFilesSectionDisplayContext
-	extends ViewFilesSectionDisplayContext {
+	extends BaseFilesSectionDisplayContext {
 
 	public ViewSpaceFilesSectionDisplayContext(
 		DepotEntryLocalService depotEntryLocalService, long groupId,
@@ -36,21 +33,7 @@ public class ViewSpaceFilesSectionDisplayContext
 			language, objectDefinitionService,
 			objectDefinitionSettingLocalService, portal);
 
-		this.groupId = groupId;
-		this.portal = portal;
-	}
-
-	@Override
-	public Map<String, Object> getEmptyState() {
-		return HashMapBuilder.<String, Object>put(
-			"description",
-			language.get(
-				httpServletRequest, "create-and-manage-files-within-this-space")
-		).put(
-			"image", "/states/cms_empty_state_files.svg"
-		).put(
-			"title", language.get(httpServletRequest, "no-files-yet")
-		).build();
+		_groupId = groupId;
 	}
 
 	@Override
@@ -58,10 +41,14 @@ public class ViewSpaceFilesSectionDisplayContext
 		return String.format(
 			"groupIds/any(g:g eq %s) and cmsSection eq 'files' and cmsRoot " +
 				"eq true",
-			groupId);
+			_groupId);
 	}
 
-	protected final long groupId;
-	protected final Portal portal;
+	@Override
+	protected String getEmptyStateDescriptionKey() {
+		return "create-and-manage-files-within-this-space";
+	}
+
+	private final long _groupId;
 
 }
